Create travis build for RMG binaries on anaconda.org
Check whatever binaries are currently on anaconda.org

Not all the unit tests get brought over to the
anaconda site packages, so it's not useful to run the full test suite.
Instead we run a few examples.

Don't use Makefile 'egX' targets for testing.
These make the 'all' target, which compiles RMG.
We want to be testing the anaconda binaries.
Instead we try to do the equivalent tests manually.

This commit merged from several by Connie Gao and Richard West.